Conflicts:
	MapView/Map/RMMapView.m
	MapView/Map/Resources/LoadingTileZoom.png
	MapView/Map/SMCalloutView
	MapView/MapView.xcodeproj/project.pbxproj
	Mapbox.podspec

+++ MapView/Map/RMQuadTree.m
@@ -522,7 +522,7 @@
 
                 _cachedClusterAnnotation = [[RMAnnotation alloc] initWithMapView:_mapView
                                                                      coordinate:clusterMarkerCoordinate
-                                                                       andTitle:[NSString stringWithFormat:@"%d", enclosedAnnotationsCount]];
+                                                                       andTitle:[NSString stringWithFormat:@"%lu", (unsigned long)enclosedAnnotationsCount]];
                 _cachedClusterAnnotation.isClusterAnnotation = YES;
                 _cachedClusterAnnotation.userInfo = self;
